requirements.txt
appnope==0.1.0
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
decorator==4.2.1
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0.2
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
futures==3.1.1
goslate==1.5.1
gunicorn==19.7.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Jinja2==2.9.6
MarkupSafe==1.0
parso==0.1.1
pep8==1.7.0
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pycrypto==2.6.1
PyDictionary==1.5.2
Pygments==2.2.0
PyJWT==1.6.0
python-http-client==3.0.0
pytz==2017.2
requests==2.18.4
sendgrid==5.3.0
simple-crypt==4.1.7
simplegeneric==0.8.1
six==1.11.0
static3==0.7.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
wikipedia==1.4.0
